third_party/blink/web_tests/http/tests/devtools/elements/styles-2/filter-matched-styles-hides-separators-expected.txt

Verifies that filtering in StylesSidebarPane hides sidebar separators.


Running: selectInitialNode

Running: onNodeSelected
element.style {
}
#third {
   font-family: times
   display: block
}
div {
   display: block
}
#second {
   font-family: helvetica
}
#first {
   font-family: arial
   display: block
}
#third::before {
   content: "uno-1"
}
#third::after {
   content: "dos-2"
   display: block
}

Running: testFilterFontFamily
Filtering styles by: font-family
[ VISIBLE ] Inherited from div#second
[ VISIBLE ] Inherited from div#first
[ HIDDEN ]  Pseudo ::before element
[ HIDDEN ]  Pseudo ::after element

Running: testContentProperty
Filtering styles by: content
[ HIDDEN ]  Inherited from div#second
[ HIDDEN ]  Inherited from div#first
[ VISIBLE ] Pseudo ::before element
[ VISIBLE ] Pseudo ::after element

Running: testDisplayProperty
Filtering styles by: display
[ HIDDEN ]  Inherited from div#second
[ VISIBLE ] Inherited from div#first
[ HIDDEN ]  Pseudo ::before element
[ VISIBLE ] Pseudo ::after element
